rename toolboxtests to toolboxlayouttests
